+++ tests/end_to_end/conftest.py
@@ -108,26 +108,6 @@ def others(addresses):
     return addresses[2:]
 
 
-@pytest_asyncio.fixture(scope="function")
-async def new_account(max_fee):
-    """
-    Return a random funded new account.
-    """
-    from kakarot_scripts.utils.kakarot import get_eoa
-    from kakarot_scripts.utils.starknet import fund_address
-    from tests.utils.helpers import generate_random_private_key
-
-    private_key = generate_random_private_key()
-    account = Wallet(
-        address=private_key.public_key.to_checksum_address(),
-        private_key=private_key,
-        # deploying an account with enough ETH to pass ~10 tx
-        starknet_contract=await get_eoa(private_key, amount=100 * max_fee / 1e18),
-    )
-    await fund_address(account.starknet_contract.address, 10)
-    return account
-
-
 @pytest_asyncio.fixture(scope="session")
 async def deployer() -> Account:
     """
